// enums/validationMessages.ts
export enum ValidationErrorMessages {
  REQUIRED = "This field is required",
  LESS_THAN_2_CHARS = "The field value must be less than 50 characters",
  LONGER_THAN_50_CHARS = "The value must be less than 50 characters long",
  LONGER_THAN_150_CHARS = "The value must be less than 150 characters long",
  LONGER_THAN_200_CHARS = "The value must be less than 200 characters long",
  UNCORRECTED_FORMAT = "Uncorrected field format",
}